Refactor expects into utilities folder
